requirements.txt
google-generativeai>=0.8.5
lancedb>=0.22.0
pydantic>=2.11.3
streamlit>=1.45.1
fastapi>=0.115.12
uvicorn>=0.34.2
python-multipart>=0.0.20
python-dotenv>=1.1.0
SQLAlchemy>=2.0.40
pytest>=8.3.5
pytest-mock>=3.0.0
pytest-cov>=6.0.0
tqdm>=4.67.1
rich>=13.9.4
numpy>=2.2.3
pandas>=2.2.3
# PDF processing dependencies
PyPDF2>=3.0.0
pdfplumber>=0.10.1
# Optional OCR dependencies
pytesseract>=0.3.10
pdf2image>=1.16.3